+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MINOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Удалены подключения плагинов yamoney-architecture-test-plugin, yamoney-build-monitoring-plugin
 ### NEXT_VERSION_DESCRIPTION_END
 ## [5.3.0]() (13-06-2019)